🚚 Move: 컴포넌트 폴더 재구성
